Electronic layout update
Replaced the wheel motors with the correct one. I will replace weapon motor. Ideal spot for weapon motor is in the back center of the robot but right now its not wide enough. Placed batteries next to each other because they do not fit end to end with the wheel motor. I will also add weights. Drew a vertical support in top view. Note, a bearing in the wall can support the ends of the motor shafts